# doc/apidoc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Khazana POS API",
    "description": "## Khazana POS Backend API\n\nAvailable modules:\n- **Auth** – Login, profile, change password\n- **Users** – User management and role assignment\n- **Sales** – Cash, credit, VAT cash, VAT credit sales\n- **Inventory** – Items, stock receive / issue / adjust\n- **Customers** – Customer management, ledger, payments\n- **Roles** – Role management\n- **Menus** – Menu items\n- **Permissions** – Role-based permissions\n- **Orders** – Sales orders (regular and VAT)\n- **Pricing** – Sale prices and cost prices\n- **Finance** – Money receive and cash purchases\n- **Packets** – Packet stock management\n- **NC Adjustment** – Non-conformance adjustments\n- **Assortment** – Assortment records\n- **Reports** – Sales, stock, item-wise, customer statements\n- **Admin** – Branches, settings, banks, audit log\n- **Dashboard** – Summary statistics\n- **Categories** – Item category management",
    "version": "1.0",
    "contact": {}
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token"
      }
    },
    "schemas": {
...
      "SaleItemDto": {
        "type": "object",
        "required": ["itemOID", "qty"],
        "properties": {
          "itemSLNum": { "type": "string", "example": "1" },
          "itemOID": { "type": "string", "format": "uuid" },
          "qty": { "type": "number", "example": 5 },
          "uom": { "type": "string", "example": "PCS" },
          "price": { "type": "number", "example": 120.00 },
          "amount": { "type": "number", "example": 600.00 },
          "sdValue": { "type": "number", "example": 0 },
          "sdAmount": { "type": "number", "example": 0 },
          "vatValue": { "type": "number", "example": 15 },
          "vatAmount": { "type": "number", "example": 90.00 },
          "discount": { "type": "number", "example": 0 },
          "netAmount": { "type": "number", "example": 690.00 }
        }
      },
      "CreateCashSaleDto": {
        "type": "object",
        "required": ["invDate", "branchId", "items"],
        "properties": {
          "invNo": { "type": "string", "example": "CS-2024-001" },
          "invDate": { "type": "string", "format": "date", "example": "2024-01-15" },
          "branchId": { "type": "integer", "example": 1 },
          "totalAmount": { "type": "number", "example": 1200.00 },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SaleItemDto" }
          }
        }
      },
      "CreateCreditSaleDto": {
        "type": "object",
        "required": ["invDate", "branchId", "clientCode", "items"],
        "properties": {
          "invNo": { "type": "string", "example": "CR-2024-001" },
          "invDate": { "type": "string", "format": "date", "example": "2024-01-15" },
          "branchId": { "type": "integer", "example": 1 },
          "clientCode": { "type": "string", "example": "CUST001" },
          "totalAmount": { "type": "number", "example": 1200.00 },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SaleItemDto" }
          }
        }
      },
      "CreateVatCashSaleDto": {
        "type": "object",
        "required": ["invDate", "branchId", "items"],
        "properties": {
          "invNo": { "type": "string", "example": "VCS-2024-001" },
          "invDate": { "type": "string", "format": "date", "example": "2024-01-15" },
          "branchId": { "type": "integer", "example": 1 },
          "totalAmount": { "type": "number", "example": 1380.00 },
          "totalVat": { "type": "number", "example": 180.00 },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SaleItemDto" }
          }
        }
      },
      "CreateVatCreditSaleDto": {
        "type": "object",
        "required": ["invDate", "branchId", "clientCode", "items"],
        "properties": {
          "invNo": { "type": "string", "example": "VCR-2024-001" },
          "invDate": { "type": "string", "format": "date", "example": "2024-01-15" },
          "branchId": { "type": "integer", "example": 1 },
          "clientCode": { "type": "string", "example": "CUST001" },
          "totalAmount": { "type": "number", "example": 1380.00 },
          "totalVat": { "type": "number", "example": 180.00 },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SaleItemDto" }
          }
        }
      },
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": { "type": "integer", "example": 100 },
          "page": { "type": "integer", "example": 1 },
          "limit": { "type": "integer", "example": 10 },
          "totalPages": { "type": "integer", "example": 10 },
          "hasNextPage": { "type": "boolean", "example": true },
          "hasPreviousPage": { "type": "boolean", "example": false }
        }
      },
...
    }
  },
  "paths": {
...
    "/sales": {
      "get": {
        "tags": ["Sales"],
        "summary": "Get all sales (filtered by type and branch)",
        "operationId": "SalesController_findAll",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "page", "in": "query", "required": false, "schema": { "type": "integer", "default": 1, "minimum": 1 }, "description": "Page number" },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 10, "minimum": 1, "maximum": 100 }, "description": "Items per page" },
          { "name": "type", "in": "query", "required": false, "schema": { "type": "string", "enum": ["cash", "credit", "vat-cash", "vat-credit"], "default": "cash" }, "description": "Sale type" },
          { "name": "branchId", "in": "query", "required": false, "schema": { "type": "integer" }, "description": "Filter by branch ID" }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of sales",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Sale retrieved successfully" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": { "type": "array", "items": { "type": "object" } },
                        "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/sales/cash": {
      "post": {
        "tags": ["Sales"],
        "summary": "Create a cash sale",
        "operationId": "SalesController_createCashSale",
        "security": [{ "access-token": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCashSaleDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "Cash sale created successfully" },
          "400": { "description": "Invalid sale data" }
        }
      }
    },
    "/sales/credit": {
      "post": {
        "tags": ["Sales"],
        "summary": "Create a credit sale",
        "operationId": "SalesController_createCreditSale",
        "security": [{ "access-token": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCreditSaleDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "Credit sale created successfully" }
        }
      }
    },
    "/sales/vat/cash": {
      "post": {
        "tags": ["Sales"],
        "summary": "Create a VAT cash sale",
        "operationId": "SalesController_createVatCashSale",
        "security": [{ "access-token": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateVatCashSaleDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "VAT cash sale created successfully" }
        }
      }
    },
    "/sales/vat/credit": {
      "post": {
        "tags": ["Sales"],
        "summary": "Create a VAT credit sale",
        "operationId": "SalesController_createVatCreditSale",
        "security": [{ "access-token": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateVatCreditSaleDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "VAT credit sale created successfully" }
        }
      }
    },
...
    "/roles": {
      "get": {
        "tags": ["Roles"],
        "summary": "Get all roles with permissions",
        "operationId": "RolesController_findAll",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "page", "in": "query", "required": false, "schema": { "type": "integer", "default": 1, "minimum": 1 }, "description": "Page number" },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 10, "minimum": 1, "maximum": 100 }, "description": "Items per page" }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of roles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Role retrieved successfully" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": { "type": "array", "items": { "type": "object" } },
                        "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Roles"],
        "summary": "Create a new role",
        "operationId": "RolesController_create",
        "security": [{ "access-token": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "type": "object" } }
          }
        },
        "responses": {
          "201": { "description": "Role created successfully" },
          "409": { "description": "Role name already exists" }
        }
      }
    },
    "/roles/{id}": {
      "get": {
        "tags": ["Roles"],
        "summary": "Get role by ID",
        "operationId": "RolesController_findOne",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "description": "Role UUID" }
        ],
        "responses": {
          "200": { "description": "Role found" },
          "404": { "description": "Role not found" }
        }
      },
      "patch": {
        "tags": ["Roles"],
        "summary": "Update role by ID",
        "operationId": "RolesController_update",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "description": "Role UUID" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "type": "object" } }
          }
        },
        "responses": {
          "200": { "description": "Role updated successfully" }
        }
      },
      "delete": {
        "tags": ["Roles"],
        "summary": "Delete role by ID",
        "operationId": "RolesController_remove",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "description": "Role UUID" }
        ],
        "responses": {
          "200": { "description": "Role deleted successfully" }
        }
      }
    },
...
    "/permissions/role/{roleId}": {
      "get": {
        "tags": ["Permissions"],
        "summary": "Get all permissions for a role",
        "operationId": "PermissionsController_findByRole",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "roleId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "description": "Role UUID" }
        ],
        "responses": {
          "200": { "description": "List of permissions for the role" }
        }
      }
    },
...
    "/pricing/prices": {
      "get": {
        "tags": ["Pricing"],
        "summary": "Get sale prices (optionally filter by item code)",
        "operationId": "PricingController_findAllPrices",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "page", "in": "query", "required": false, "schema": { "type": "integer", "default": 1, "minimum": 1 }, "description": "Page number" },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 10, "minimum": 1, "maximum": 100 }, "description": "Items per page" },
          { "name": "itemCode", "in": "query", "required": false, "schema": { "type": "string" }, "description": "Filter by item code" }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of sale prices",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Price retrieved successfully" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": { "type": "array", "items": { "type": "object" } },
                        "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Pricing"],
        "summary": "Create a new sale price",
        "operationId": "PricingController_createPrice",
        "security": [{ "access-token": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "type": "object" } }
          }
        },
        "responses": {
          "201": { "description": "Sale price created successfully" }
        }
      }
    },
    "/pricing/prices/{id}": {
      "patch": {
        "tags": ["Pricing"],
        "summary": "Update a sale price by ID",
        "operationId": "PricingController_updatePrice",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "description": "Price record UUID" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "type": "object" } }
          }
        },
        "responses": {
          "200": { "description": "Sale price updated successfully" },
          "404": { "description": "Price record not found" }
        }
      }
    },
    "/pricing/prices/current": {
      "get": {
        "tags": ["Pricing"],
        "summary": "Get current active price for an item on a given date",
        "operationId": "PricingController_getCurrentPrice",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "itemCode", "in": "query", "required": true, "schema": { "type": "string" }, "description": "Item code" },
          { "name": "date", "in": "query", "required": false, "schema": { "type": "string", "format": "date" }, "description": "Date to check price (ISO 8601, defaults to today)" }
        ],
        "responses": {
          "200": { "description": "Current price record" }
        }
      }
    },
    "/pricing/cost-prices": {
      "get": {
        "tags": ["Pricing"],
        "summary": "Get cost prices (optionally filter by item code)",
        "operationId": "PricingController_findAllCostPrices",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "page", "in": "query", "required": false, "schema": { "type": "integer", "default": 1, "minimum": 1 }, "description": "Page number" },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 10, "minimum": 1, "maximum": 100 }, "description": "Items per page" },
          { "name": "itemCode", "in": "query", "required": false, "schema": { "type": "string" }, "description": "Filter by item code" }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of cost prices",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Cost Price retrieved successfully" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": { "type": "array", "items": { "type": "object" } },
                        "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Pricing"],
        "summary": "Create a new cost price",
        "operationId": "PricingController_createCostPrice",
        "security": [{ "access-token": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "type": "object" } }
          }
        },
        "responses": {
          "201": { "description": "Cost price created successfully" }
        }
      }
    },
    "/pricing/cost-prices/{id}": {
      "patch": {
        "tags": ["Pricing"],
        "summary": "Update a cost price by ID",
        "operationId": "PricingController_updateCostPrice",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "description": "Cost price record UUID" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "type": "object" } }
          }
        },
        "responses": {
          "200": { "description": "Cost price updated successfully" },
          "404": { "description": "Cost price record not found" }
        }
      }
    },
...
  },
...
}